Keda_+_CPU_Based_Scheduling _with_node_affinity_+_K3s/requirements-local.txt
fastapi==0.115.0
uvicorn==0.30.6
grpcio==1.64.1
grpcio-tools==1.64.1
protobuf==5.27.3
prometheus-api-client==0.5.4
httpx==0.27.2
requests==2.32.3
kubernetes==30.1.0
